pyproject.toml
[tool.poetry]
name = "mlali-agents"
version = "0.1.0"
description = "Agents untuk riset aplikasi mlali"
authors = ["Resika Arthana <resika.arthana@gmail.com>"]
license = "Edukasi"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
langchain = "^0.3.10"
langgraph = "^0.2.56"
langchain-core = "^0.3.22"
langchain-openai = "^0.2.11"
langchain-community = "^0.3.10"
typing-extensions = "^4.12.2"
pypdf = "^5.1.0"
python-multipart = "^0.0.19"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
faiss-cpu = "1.8.0"
fastapi = "^0.115.6"
uvicorn = "^0.32.1"
streamlit = "^1.40.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
